httplib: Move httplib/serve/data to ../serve/http/data
