+++ cmd/serve/http/data/assets_generate.go
